+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+<<<<<<< HEAD
+=======
+## [v1.2.1] - 2023-08-22
+
+### Added
+* Added acceptance tests
+
+### Changed
+* Update unit tests
+* Removed Centos 8 support
+
+>>>>>>> 302e9ad (update unit tests, pdk3.0 and add acceptance tests)
 ## [v1.2.0] - 2023-08-14
 
 ### Changed